refactor: split main.go into balance.go and import.go
Split the growing main.go into three files:
- main.go: app setup, route registration, record hooks
- balance.go: balance worker, enqueueBalance, calculateBalance
- import.go: import/revert types, helpers, and handlers

Also fix pb-server.ts build hash to include all .go files (not just
main.go) so changes to any Go file trigger a rebuild.

Remove stale pb_hooks eslint ignore.
