+++ eslint.config.js
@@ -11,7 +11,7 @@ import svelteConfig from './svelte.config.js';
 const gitignorePath = fileURLToPath(new URL('./.gitignore', import.meta.url));
 
 export default ts.config(
-	{ ignores: ['pocketbase/pb_migrations/**', 'pocketbase/pb_hooks/**'] },
+	{ ignores: ['pocketbase/pb_migrations/**'] },
 	includeIgnoreFile(gitignorePath),
 	js.configs.recommended,
 	...ts.configs.recommended,
